feat: add styles and preloader
