+++ src/components/card/card.module.css
@@ -1,19 +1,15 @@
 .card {
   display: flex;
   flex-direction: column;
-  margin: 1.2rem;
+  margin: 1rem;
   border-radius: 4px;
-  /* background: rgb(252, 252, 252); */
-  /* border: 1px solid rgba(179, 179, 181, 0.35); */
-  background: rgb(240, 240, 234);
-  border: 1px solid rgba(164,139,113, 0.4);
+  background: rgb(252, 252, 252);
+  border: 1px solid rgba(109, 102, 94, 0.1);
   min-height: 300px;
 }
 
 .card:hover{
-  /* -webkit-box-shadow: 0px 2px 8px 2px rgba(128, 129, 150, 0.31); */
-  /* -moz-box-shadow: 0px 2px 8px 2px rgba(128, 129, 150, 0.31); */
-  box-shadow: 0px 2px 8px 2px rgba(164,139,113, 0.4);
+  box-shadow: 0px 2px 8px 2px rgba(109, 102, 94, 0.3);
 }
 
 .image > img {
