+++ src/components/catModal/catmodal.module.css
@@ -2,23 +2,17 @@
 .image > img {
   display: block;
   width: 100%;
+  height: 100%;
   object-position: top;
   object-fit: cover;
   aspect-ratio: 1.5/1;
   border-radius: 8px 8px 0px 0px;
 }
 
-.info {
-  margin: 0.6rem;
-  line-height: 1.5;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
 .overlay {
   position: fixed;
   width: 100%;
+  max-width: 100%;
   height: 100vh;
   top: 0;
   left: 0;
@@ -26,6 +20,18 @@
   transition: all 0.3s ease;
   z-index: 3;
   backdrop-filter: blur(8px);
+  padding: 1rem;
+}
+
+.info {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 37rem;
+  max-width: 100%;
+  height: 43rem;
+  transform: translate(-50%, -50%);
+  padding: 1rem;
 }
 
 .modal {
@@ -33,10 +39,12 @@
   top: 50%;
   left: 50%;
   width: 35rem;
+  max-width: calc(100% - 2rem);
+  height: 43rem;
   transform: translate(-50%, -50%);
   border-radius: 8px;
-  box-shadow: 0px 4px 8px 4px rgba(116, 95, 73, 0.5);
-  background: rgb(116, 95, 73);
+  box-shadow: 0px 4px 8px 4px rgba(109, 102, 94, 0.5);
+  background: rgb(109, 102, 94);
   color: white;
   transition: all 0.3s ease;
   z-index: 4;
@@ -45,6 +53,7 @@
 }
 
 .container {
+  width: 100%;
 text-align: center;
 color: white;
 padding: 1rem;
@@ -88,4 +97,5 @@ padding: 1rem;
   color: white;
   line-height: 1;
   font-size: 20px;
+  border: none;
 }
