+++ src/components/formData/FormData.test.tsx
@@ -1,8 +1,9 @@
-import { render, screen } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import { describe, test, vi } from 'vitest';
 import userEvent from '@testing-library/user-event';
 import { FormData } from './FormData';
 import { Countries } from '../../types/types';
+import { InputAgree } from './InputAgree';
 
 describe('FormData', () => {
   beforeAll(() => {
@@ -30,14 +31,30 @@ describe('FormData', () => {
       selector: 'input[type="file"]',
     });
     expect(fileInput).toBeInTheDocument();
-
-    test('remove values', async () => {
-      await userEvent.type(nameInput, 'test');
-      await userEvent.selectOptions(countryInput, Countries.France);
-      userEvent.click(screen.getByRole('button', { name: 'Submit' }));
-
-      expect(nameInput.value).toBe('');
-      expect(countryInput.value).toBe('');
-    });
   });
+  // test('remove values', async () => {
+  //   const showCardsValues = vi.fn();
+  //   const setShowModal = vi.fn();
+  //   render(<FormData showCardsValues={showCardsValues} showModal={setShowModal} />);
+  //   const nameInput = screen.getByLabelText<HTMLInputElement>(/name/i);
+  //   await userEvent.type(nameInput, 'test');
+
+  //   const countryInput = screen.getByLabelText<HTMLSelectElement>(/country/i);
+  //   await userEvent.selectOptions(countryInput, Countries.France);
+
+  //   const button = screen.getByRole('button', { name: 'Submit' });
+  //   await userEvent.click(button);
+  //   console.debug;
+  //   expect(nameInput.value).toEqual('');
+  //   expect(countryInput.value).toBe('');
+  // });
+  // test('checkbox', async () => {
+  //   const { container } = render(<InputAgree />);
+  //   console.debug
+  //   const checkbox = container.firstChild;
+
+  //   expect(checkbox).not.toBeChecked();
+  //   await userEvent.click(checkbox);
+  //   expect(checkbox).toBeChecked();
+  // });
 });
