+++ src/components/header/header.module.css
@@ -2,12 +2,12 @@
   display: flex;
   justify-content: space-around;
   flex: 0 0 auto;
-  background: rgb(86, 123, 149);
+  background: rgb(132, 163, 185);
 }
 
 .title {
   font-weight: 800;
   text-transform: uppercase;
   padding: 1rem;
-  color: rgb(25, 42, 54);;
+  color: rgb(25, 42, 54);
 }
