+++ src/components/nav/nav.module.css
@@ -14,6 +14,6 @@
   color: white;
   text-decoration: none;
   padding: 1rem;
-  background-color: rgb(63, 100, 127);
+  background-color: rgb(106, 141, 167);
   cursor: default;
 }
