+++ src/components/preloader/preloader.module.css
@@ -1,0 +1,6 @@
+.loader {
+  display: block;
+  margin: 0 auto;
+  width: 64px;
+  height: 64px;
+}
