+++ src/components/search/search.module.css
@@ -2,7 +2,7 @@
   padding: 1rem;
   display: flex;
   justify-content: center;
-  margin: 0 auto;
+  margin: 1rem auto 0;
 }
 
 .label-search{
@@ -19,7 +19,7 @@
   padding-left: 10px;
   outline: none;
   border-radius: 8px 0px 0px 8px;
-  border: 1px solid rgba(164,139,113, 0.4);
+  border: 1px solid rgba(109, 102, 94, 0.1);
   background: rgb(252, 252, 252);
   color: rgb(21, 21, 40);
   position: relative;
@@ -42,27 +42,12 @@
 }
 
 .button-search {
-  color: rgb(63, 100, 127);;
+  color: rgb(63, 100, 127);
   background: rgb(252, 252, 252);
   border-radius: 0px 8px 8px 0px;
-  border: 1px solid rgba(164,139,113, 0.4);
+  border: 1px solid rgba(109, 102, 94, 0.1);
   padding: 10px 30px;
   cursor: pointer;
   text-transform: uppercase;
-  font-weight: 800;
-}
-
-.material-symbols-outlined {
-  z-index: 2;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 25px;
-  height: 25px;
-  transform: translate(50%, 50%);
-  font-variation-settings:
-  'FILL' 0,
-  'wght' 400,
-  'GRAD' 0,
-  'opsz' 48
+  font-weight: 700;
 }
