+++ src/index.css
@@ -17,7 +17,7 @@ body {
   padding: 0;
   font-size: 16px;
   color: rgb(21, 21, 40);
-  background: rgb(242,239,220);
+  background: rgba(255,190,122, 0.4);
   display: flex;
   flex-direction: column;
   min-height: 100vh;
